Address Martin's feedback
+ Fix a warning on iMessage example app that was happening to the AccentColor missing from the Assets catalogue
