+++ Examples/YouTubeClone/AppDelegate.swift
@@ -15,7 +15,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
         let window = UIWindow(frame: UIScreen.main.bounds)
-        window.rootViewController = SplashViewController { [unowned window] in
+        window.rootViewController = SplashViewController(
+            userInfo: .init(id: "sagar"),
+            token: .development(userId: "sagar")
+        ) { [unowned window] in
             window.rootViewController = UINavigationController(
                 rootViewController: YTLiveVideoViewController()
             )
@@ -31,48 +34,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
 
         return true
-    }
-}
-
-private final class SplashViewController: UIViewController {
-    private let userInfo: UserInfo
-    private let token: Token
-    private let completionHandler: () -> Void
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-    
-    init(
-        userInfo: UserInfo = .init(id: "sagar"),
-        token: Token = .development(userId: "sagar"),
-        completionHandler: @escaping () -> Void
-    ) {
-        self.userInfo = userInfo
-        self.token = token
-        self.completionHandler = completionHandler
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .systemBackground
-        let activityIndicatorView = UIActivityIndicatorView(style: .large)
-        activityIndicatorView.startAnimating()
-        activityIndicatorView.center = view.center
-        view.addSubview(activityIndicatorView)
-        
-        ChatClient.shared.connectUser(
-            userInfo: userInfo,
-            token: token
-        ) { [weak self] error in
-            if let error {
-                fatalError("Failed to connect user.(\(error.localizedDescription)")
-            } else {
-                DispatchQueue.main.async {
-                    self?.completionHandler()
-                }
-            }
-        }
     }
 }
